Forward inbox requests to Lemmy
