+++ docker/nginx.conf
@@ -57,16 +57,7 @@ server {
 
     # lemmybb frontend
     location / {
-        set $proxpass "http://0.0.0.0:8701";
-        # federation requests go to backend
-        if ($http_accept ~ "^application/.*$") {
-            set $proxpass "http://0.0.0.0:8703";
-        }
-        # fonts have similar accept header but need to be handled by frontend
-        if ($http_accept ~ "^application/font.*$") {
-            set $proxpass "http://0.0.0.0:8701";
-        }
-        proxy_pass $proxpass;
+        proxy_pass "http://0.0.0.0:8701";
         rewrite ^(.+)/+$ $1 permanent;
     }
