+++ src/main.rs
@@ -20,9 +20,9 @@ use crate::{
     routes::{
         comment::*,
         community::*,
+        federation::*,
         post::*,
         private_message::*,
-        redirects::*,
         site::*,
         user::*,
     },
@@ -114,7 +114,8 @@ fn init_rocket() -> Result<Rocket<Build>, Error> {
                 do_report,
                 edit_profile,
                 do_edit_profile,
-                community_list
+                community_list,
+                inboxes
             ],
         )
         .mount("/assets", FileServer::from(relative!("assets"))))
